requirements.txt
asgiref==3.5.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Markdown==3.3.7
mysqlclient==2.1.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
django-rest-swagger
pyyaml
django-cors-headers
django-environ
Pillow
